agentics/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9
PyMuPDF==1.24.9
easyocr==1.7.1
numpy>=2.1,<3
Pillow==10.4.0
pytesseract==0.3.10
PyPDF2==3.0.1
pdfminer.six==20231228
requests==2.32.3
pdf2image==1.17.0
google-generativeai==0.7.2